Add a create() function to simplify internals
The create() function takes a builder function that takes two callbacks,
instead of one: a fulfill() and reject() callback, for fulfilling or
rejecting the created Future. This is similar to how most Promise
libraries do this. It's not intended for broad usage -- the make()
function interoperates with Node better -- but it helps simplify the
internals by reducing some code duplication, and helps interoperate with
promise libraries more simply.