Fix collectAll and collectAllByTimestamp
Previously, it was possible for both to error out, since they passed
along the error param as if it were an error for the collection
function. However, since both of them collect errors into Result tuples,
neither collectors should ever error: the errors are part of the
returned values, not themselves an error in the function.

Also, collectAllByTimestamp mistakenly used push() rather than setting
values in the index. Since the arrays are pre-allocated to a fixed size,
this resulted in the first elements of the arrays being left undefined.